requirements.txt
django 
djangorestframework 
Pillow
django-extensions
ipython